Change detector threshold parameter from float to Double
Change threshold parameter type in BaseDetector from float to Double

+++ modules/watsonx-ai/src/main/java/com/ibm/watsonx/ai/detection/detector/BaseDetector.java
@@ -62,7 +62,7 @@ public abstract class BaseDetector {
             properties = new HashMap<>();
         }
 
-        public T threshold(float threshold) {
+        public T threshold(Double threshold) {
             return addProperty("threshold", threshold);
         }
